Change Config Dependency Injection Change to loading of Modules by
