+++ src/Athavar.FFXIV.Plugin.Macro/DependencyInjection.cs
@@ -2,6 +2,7 @@
 // Copyright (c) Athavar. All rights reserved.
 // Licensed under the GPL-3.0 license. See LICENSE file in the project root for full license information.
 // </copyright>
+
 namespace Athavar.FFXIV.Plugin.Macro;
 
 using Athavar.FFXIV.Plugin.Macro.Managers;
@@ -14,7 +15,6 @@ public static class DependencyInjection
         services.AddSingleton<ConditionCheck>();
         services.AddSingleton<MacroManager>();
         services.AddSingleton<MacroHelpWindow>();
-        MacroConfiguration.AddToDependencyInjection(services);
 
         return services;
     }
