+++ src/Athavar.FFXIV.Plugin.Yes/DependencyInjection.cs
@@ -2,6 +2,7 @@
 // Copyright (c) Athavar. All rights reserved.
 // Licensed under the GPL-3.0 license. See LICENSE file in the project root for full license information.
 // </copyright>
+
 namespace Athavar.FFXIV.Plugin.Yes;
 
 using Microsoft.Extensions.DependencyInjection;
@@ -11,7 +12,6 @@ public static class DependencyInjection
     public static IServiceCollection AddYesModule(this IServiceCollection services)
     {
         services.AddSingleton<ZoneListWindow>();
-        YesConfiguration.AddToDependencyInjection(services);
 
         return services;
     }
